move shop dialogs into their own file
